remove unused comps and restructure app
